feat(person): enforce unique CPF constraint and add duplicate resourc…

+++ build/http/controllers/posts/delete-posts.js
@@ -69,6 +69,9 @@ Posts = __decorateClass([
   (0, import_typeorm.Entity)({ name: "posts" })
 ], Posts);
 
+// src/repositories/typeorm/posts.repository.ts
+var import_typeorm6 = require("typeorm");
+
 // src/entities/person.entity.ts
 var import_typeorm4 = require("typeorm");
 
@@ -252,6 +255,23 @@ var UpdateUserRoleToUppercase1773790761895 = class {
   }
 };
 
+// src/lib/typeorm/migrations/1777408444519-AlterTablePersonUniqueCpf.ts
+var AlterTablePersonUniqueCpf1777408444519 = class {
+  async up(queryRunner) {
+    await queryRunner.query(
+      `ALTER TABLE person 
+        ADD CONSTRAINT person_unique_cpf UNIQUE (cpf)`
+    );
+  }
+  async down(queryRunner) {
+    await queryRunner.query(
+      `ALTER TABLE person
+        DROP CONSTRAINT IF EXISTS person_unique_cpf
+        `
+    );
+  }
+};
+
 // src/lib/typeorm/typeorm.ts
 var appDataSource = new import_typeorm5.DataSource({
   type: "postgres",
@@ -263,7 +283,8 @@ var appDataSource = new import_typeorm5.DataSource({
   entities: [Posts, User, Person, Address],
   migrations: [
     UserAddRole1773452300096,
-    UpdateUserRoleToUppercase1773790761895
+    UpdateUserRoleToUppercase1773790761895,
+    AlterTablePersonUniqueCpf1777408444519
   ],
   logging: env.NODE_ENV === "development"
 });
@@ -291,6 +312,16 @@ var PostsRepository = class {
       }
     });
   }
+  async search(query) {
+    return this.repository.find({
+      where: [
+        { title: (0, import_typeorm6.ILike)(`%${query}%`) },
+        // Busca no título
+        { content: (0, import_typeorm6.ILike)(`%${query}%`) }
+        // Busca no conteúdo
+      ]
+    });
+  }
   async create(posts) {
     return this.repository.save(posts);
   }
@@ -311,15 +342,25 @@ var ResourceNotFoundError = class extends Error {
   }
 };
 
+// src/useCases/errors/UnauthorizedError.ts
+var UnauthorizedError = class extends Error {
+  constructor() {
+    super("Unauthorized");
+  }
+};
+
 // src/useCases/delete-posts.ts
 var DeletePostsUseCase = class {
   constructor(postsRepository) {
     this.postsRepository = postsRepository;
   }
-  async execute(id) {
+  async execute(id, role) {
     const post = await this.postsRepository.findById(id);
     if (!post) {
       throw new ResourceNotFoundError();
+    }
+    if (role !== "PROFESSOR" /* PROFESSOR */) {
+      throw new UnauthorizedError();
     }
     return this.postsRepository.delete(id);
   }
@@ -334,14 +375,15 @@ function makeDeletePostUseCase() {
 
 // src/http/controllers/posts/delete-posts.ts
 var import_zod2 = require("zod");
-async function deletePost(req, res) {
+async function deletePost(req, reply) {
   const registerParamsSchema = import_zod2.z.object({
     id: import_zod2.z.coerce.number()
   });
   const { id } = registerParamsSchema.parse(req.params);
+  const user = req.user;
   const deletepostUseCase = makeDeletePostUseCase();
-  await deletepostUseCase.execute(id);
-  return res.status(204).send();
+  await deletepostUseCase.execute(id, user.role);
+  return reply.status(204).send();
 }
 // Annotate the CommonJS export names for ESM import in node:
 0 && (module.exports = {
